Remove dead constraints and test
