+++ pintc/tests/interfaces/simple.pnt
@@ -52,10 +52,6 @@ predicate Simple {
     state x = FooInstance::storage::x;
     state y = FooInstance::storage::y;
     state z = BarInstance::storage::z;
-    constraint x == 0;
-    constraint y == 0;
-    constraint z == 0x0000000000000000000000000000000000000000000000000000000000000000;
-
 
     predicate FooI1Instance1 = FooInstance::I1(addr3);
     predicate FooI1Instance2 = FooInstance::I1(addr3);
@@ -128,9 +124,6 @@ predicate Simple {
 //     state ::x = ::FooInstance::storage::x;
 //     state ::y = ::FooInstance::storage::y;
 //     state ::z = ::BarInstance::storage::z;
-//     constraint (::x == 0);
-//     constraint (::y == 0);
-//     constraint (::z == 0x0000000000000000000000000000000000000000000000000000000000000000);
 //     constraint (::x1 == ::FooI1Instance1::x1);
 //     constraint (::x2 == ::BarI2Instance2::x2);
 //     constraint (::x3 == ::BarI2Instance3::x3);
@@ -196,9 +189,6 @@ predicate Simple {
 //     state ::x: int = __storage_get_extern(::addr1, {0});
 //     state ::y: int = __storage_get_extern(::addr1, {1});
 //     state ::z: b256 = __storage_get_extern(::addr2, {1});
-//     constraint (::x == 0);
-//     constraint (::y == 0);
-//     constraint (::z == 0x0000000000000000000000000000000000000000000000000000000000000000);
 //     constraint (::x1 == __transient({0}, 1, __::FooI1Instance1_pathway));
 //     constraint (::x2 == __transient({1}, 1, __::BarI2Instance2_pathway));
 //     constraint (::x3 == __transient({2}, 1, __::BarI2Instance3_pathway));
