+++ pintc/tests/storage/external_storage/main.pnt
@@ -38,19 +38,6 @@ predicate Foo {
     state external_x_2 = ERC20_2_instance::storage::x;
     state external_y_2 = ERC20_2_instance::storage::y;
     state external_map_2 = ERC20_2_instance::storage::map[3][4];
-
-    constraint internal_x == 0;
-    constraint internal_y == (1 == 1);
-    constraint internal_map == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint external_x_0 == 0;
-    constraint external_y_0 == (1 == 1);
-    constraint external_map_0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint external_x_1 == 0;
-    constraint external_y_1 == (1 == 1);
-    constraint external_map_1 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint external_x_2 == 0;
-    constraint external_y_2 == (1 == 1);
-    constraint external_map_2 == 0x0000000000000000000000000000000000000000000000000000000000000000;
 }
 
 predicate Bar {
@@ -76,19 +63,6 @@ predicate Bar {
     state external_x_2 = ERC20_2_instance::storage::x;
     state external_y_2 = ERC20_2_instance::storage::y;
     state external_map_2 = ERC20_2_instance::storage::map[3][4];
-
-    constraint internal_x == 0;
-    constraint internal_y == (1 == 1);
-    constraint internal_map == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint external_x_0 == 0;
-    constraint external_y_0 == (1 == 1);
-    constraint external_map_0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint external_x_1 == 0;
-    constraint external_y_1 == (1 == 1);
-    constraint external_map_1 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint external_x_2 == 0;
-    constraint external_y_2 == (1 == 1);
-    constraint external_map_2 == 0x0000000000000000000000000000000000000000000000000000000000000000;
 }
 
 // parsed <<<
@@ -135,18 +109,6 @@ predicate Bar {
 //     state ::external_x_2 = ::ERC20_2_instance::storage::x;
 //     state ::external_y_2 = ::ERC20_2_instance::storage::y;
 //     state ::external_map_2 = ::ERC20_2_instance::storage::map[3][4];
-//     constraint (::internal_x == 0);
-//     constraint (::internal_y == (1 == 1));
-//     constraint (::internal_map == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_0 == 0);
-//     constraint (::external_y_0 == (1 == 1));
-//     constraint (::external_map_0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_1 == 0);
-//     constraint (::external_y_1 == (1 == 1));
-//     constraint (::external_map_1 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_2 == 0);
-//     constraint (::external_y_2 == (1 == 1));
-//     constraint (::external_map_2 == 0x0000000000000000000000000000000000000000000000000000000000000000);
 // }
 //
 // predicate ::Bar {
@@ -165,18 +127,6 @@ predicate Bar {
 //     state ::external_x_2 = ::ERC20_2_instance::storage::x;
 //     state ::external_y_2 = ::ERC20_2_instance::storage::y;
 //     state ::external_map_2 = ::ERC20_2_instance::storage::map[3][4];
-//     constraint (::internal_x == 0);
-//     constraint (::internal_y == (1 == 1));
-//     constraint (::internal_map == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_0 == 0);
-//     constraint (::external_y_0 == (1 == 1));
-//     constraint (::external_map_0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_1 == 0);
-//     constraint (::external_y_1 == (1 == 1));
-//     constraint (::external_map_1 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_2 == 0);
-//     constraint (::external_y_2 == (1 == 1));
-//     constraint (::external_map_2 == 0x0000000000000000000000000000000000000000000000000000000000000000);
 // }
 // >>>
 
@@ -224,18 +174,6 @@ predicate Bar {
 //     state ::external_x_2: int = __storage_get_extern(0x2222222222222222222222222222222222222222222222222222222222222222, {0});
 //     state ::external_y_2: bool = __storage_get_extern(0x2222222222222222222222222222222222222222222222222222222222222222, {1});
 //     state ::external_map_2: b256 = __storage_get_extern(0x2222222222222222222222222222222222222222222222222222222222222222, {2, 3, 4});
-//     constraint (::internal_x == 0);
-//     constraint (::internal_y == (1 == 1));
-//     constraint (::internal_map == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_0 == 0);
-//     constraint (::external_y_0 == (1 == 1));
-//     constraint (::external_map_0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_1 == 0);
-//     constraint (::external_y_1 == (1 == 1));
-//     constraint (::external_map_1 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_2 == 0);
-//     constraint (::external_y_2 == (1 == 1));
-//     constraint (::external_map_2 == 0x0000000000000000000000000000000000000000000000000000000000000000);
 //     constraint __eq_set(__mut_keys(), {0});
 // }
 //
@@ -255,18 +193,6 @@ predicate Bar {
 //     state ::external_x_2: int = __storage_get_extern(0x2222222222222222222222222222222222222222222222222222222222222222, {0});
 //     state ::external_y_2: bool = __storage_get_extern(0x2222222222222222222222222222222222222222222222222222222222222222, {1});
 //     state ::external_map_2: b256 = __storage_get_extern(0x2222222222222222222222222222222222222222222222222222222222222222, {2, 3, 4});
-//     constraint (::internal_x == 0);
-//     constraint (::internal_y == (1 == 1));
-//     constraint (::internal_map == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_0 == 0);
-//     constraint (::external_y_0 == (1 == 1));
-//     constraint (::external_map_0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_1 == 0);
-//     constraint (::external_y_1 == (1 == 1));
-//     constraint (::external_map_1 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::external_x_2 == 0);
-//     constraint (::external_y_2 == (1 == 1));
-//     constraint (::external_map_2 == 0x0000000000000000000000000000000000000000000000000000000000000000);
 //     constraint __eq_set(__mut_keys(), {0});
 // }
 // >>>
