+++ pintc/tests/storage/mut_storage.pnt
@@ -21,16 +21,6 @@ predicate Foo {
 
     // Should not show up in the list of mutable keys
     state map_43 = storage::map[42].1;
-
-    constraint x == 0;
-    constraint b == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint y_2 == { 0x0000000000000000000000000000000000000000000000000000000000000000, 0 };
-    constraint z == [[0, 0, 0], [0, 0, 0]];
-    constraint w_1 == [true, false, true];
-    constraint v_3 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint map_69 == { 0x0000000000000000000000000000000000000000000000000000000000000000, 0 };
-    constraint map_42_1 == 0;
-    constraint map_43 == 0;
 }
 
 // parsed <<<
@@ -55,15 +45,6 @@ predicate Foo {
 //     state ::map_42_1 = mut storage::map[42].1;
 //     state ::vec_len = __vec_len(mut storage::vec);
 //     state ::map_43 = storage::map[42].1;
-//     constraint (::x == 0);
-//     constraint (::b == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::y_2 == {0x0000000000000000000000000000000000000000000000000000000000000000, 0});
-//     constraint (::z == [[0, 0, 0], [0, 0, 0]]);
-//     constraint (::w_1 == [true, false, true]);
-//     constraint (::v_3 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::map_69 == {0x0000000000000000000000000000000000000000000000000000000000000000, 0});
-//     constraint (::map_42_1 == 0);
-//     constraint (::map_43 == 0);
 // }
 // >>>
 
@@ -87,18 +68,8 @@ predicate Foo {
 //     state ::v_3: b256 = __storage_get({5, 3});
 //     state ::map_69: {b256, int} = __storage_get({6, 69, 0});
 //     state ::map_42_1: int = __storage_get({6, 42, (0 + 1)});
+//     state ::vec_len: int = __storage_get({5});
 //     state ::map_43: int = __storage_get({6, 42, (0 + 1)});
-//     state __vec_len: int = __storage_get({5});
-//     constraint (::x == 0);
-//     constraint (::b == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::y_2 == {0x0000000000000000000000000000000000000000000000000000000000000000, 0});
-//     constraint (::z == [[0, 0, 0], [0, 0, 0]]);
-//     constraint (::w_1 == [true, false, true]);
-//     constraint (::v_3 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::map_69 == {0x0000000000000000000000000000000000000000000000000000000000000000, 0});
-//     constraint (::map_42_1 == 0);
-//     constraint (::map_43 == 0);
-//     constraint (3 < __vec_len);
 //     constraint __eq_set(__mut_keys(), {{0}, 1, {1}, 1, {2, (0 + 2)}, 2, {2, ((0 + 2) + 1)}, 2, {3, 0}, 2, {3, (0 + 1)}, 2, {3, (0 + 2)}, 2, {3, (0 + 3)}, 2, {3, (0 + 4)}, 2, {3, (0 + 5)}, 2, {4, (0 + (1 * 3))}, 2, {4, ((0 + (1 * 3)) + 1)}, 2, {4, ((0 + (1 * 3)) + 2)}, 2, {5, 3}, 2, {6, 69, 0}, 3, {6, 69, (0 + 1)}, 3, {6, 42, (0 + 1)}, 3, {5}, 1, 54});
 // }
 // >>>
