+++ pintc/tests/storage/storage_tuple.pnt
@@ -54,37 +54,6 @@ predicate Bar {
     state foo_map_to_tuples_69_0 = FooInstance::storage::map_to_tuples[69].0;
     state foo_map_to_tuples_69_1_0 = FooInstance::storage::map_to_tuples[69].1.0;
     state foo_map_to_tuples_69_1_1 = FooInstance::storage::map_to_tuples[69].1.1;
-
-    constraint u == { 0x0000000000000000000000000000000000000000000000000000000000000000, 0 };
-    constraint u0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint u1 == 0;
-    constraint t == { 0x0000000000000000000000000000000000000000000000000000000000000000, { 0, 0 } };
-    constraint t0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint t10 == 0;
-    constraint t11 == 0;
-    constraint w == { addr: 0x0000000000000000000000000000000000000000000000000000000000000000 , inner: { 0, 0 } };
-    constraint w0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint w10 == 0;
-    constraint w11 == 0;
-    constraint map_to_tuples_69 == { 0x0000000000000000000000000000000000000000000000000000000000000000, { 0, 0 } };
-    constraint map_to_tuples_69_0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint map_to_tuples_69_1_0 == 0;
-    constraint map_to_tuples_69_1_1 == 0;
-    constraint foo_u == { 0x0000000000000000000000000000000000000000000000000000000000000000, 0 };
-    constraint foo_u0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint foo_u1 == 0;
-    constraint foo_t == { 0x0000000000000000000000000000000000000000000000000000000000000000, { 0, 0 } };
-    constraint foo_t0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint foo_t10 == 0;
-    constraint foo_t11 == 0;
-    constraint foo_w == { addr: 0x0000000000000000000000000000000000000000000000000000000000000000, inner: { x: 0, 0 } };
-    constraint foo_w0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint foo_w10 == 0;
-    constraint foo_w11 == 0;
-    constraint foo_map_to_tuples_69 == { 0x0000000000000000000000000000000000000000000000000000000000000000, { 0, 0 } };
-    constraint foo_map_to_tuples_69_0 == 0x0000000000000000000000000000000000000000000000000000000000000000;
-    constraint foo_map_to_tuples_69_1_0 == 0;
-    constraint foo_map_to_tuples_69_1_1 == 0;
 }
 
 // parsed <<<
@@ -135,36 +104,6 @@ predicate Bar {
 //     state ::foo_map_to_tuples_69_0 = ::FooInstance::storage::map_to_tuples[69].0;
 //     state ::foo_map_to_tuples_69_1_0 = ::FooInstance::storage::map_to_tuples[69].1.0;
 //     state ::foo_map_to_tuples_69_1_1 = ::FooInstance::storage::map_to_tuples[69].1.1;
-//     constraint (::u == {0x0000000000000000000000000000000000000000000000000000000000000000, 0});
-//     constraint (::u0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::u1 == 0);
-//     constraint (::t == {0x0000000000000000000000000000000000000000000000000000000000000000, {0, 0}});
-//     constraint (::t0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::t10 == 0);
-//     constraint (::t11 == 0);
-//     constraint (::w == {addr: 0x0000000000000000000000000000000000000000000000000000000000000000, inner: {0, 0}});
-//     constraint (::w0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::w10 == 0);
-//     constraint (::w11 == 0);
-//     constraint (::map_to_tuples_69 == {0x0000000000000000000000000000000000000000000000000000000000000000, {0, 0}});
-//     constraint (::map_to_tuples_69_0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::map_to_tuples_69_1_0 == 0);
-//     constraint (::map_to_tuples_69_1_1 == 0);
-//     constraint (::foo_u == {0x0000000000000000000000000000000000000000000000000000000000000000, 0});
-//     constraint (::foo_u0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::foo_u1 == 0);
-//     constraint (::foo_t == {0x0000000000000000000000000000000000000000000000000000000000000000, {0, 0}});
-//     constraint (::foo_t0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::foo_t10 == 0);
-//     constraint (::foo_t11 == 0);
-//     constraint (::foo_w == {addr: 0x0000000000000000000000000000000000000000000000000000000000000000, inner: {x: 0, 0}});
-//     constraint (::foo_w0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::foo_w10 == 0);
-//     constraint (::foo_w11 == 0);
-//     constraint (::foo_map_to_tuples_69 == {0x0000000000000000000000000000000000000000000000000000000000000000, {0, 0}});
-//     constraint (::foo_map_to_tuples_69_0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::foo_map_to_tuples_69_1_0 == 0);
-//     constraint (::foo_map_to_tuples_69_1_1 == 0);
 // }
 // >>>
 
@@ -216,36 +155,6 @@ predicate Bar {
 //     state ::foo_map_to_tuples_69_0: b256 = __storage_get_extern(0x1111111111111111111111111111111111111111111111111111111111111111, {3, 69, (0 + 0)});
 //     state ::foo_map_to_tuples_69_1_0: int = __storage_get_extern(0x1111111111111111111111111111111111111111111111111111111111111111, {3, 69, ((0 + 1) + 0)});
 //     state ::foo_map_to_tuples_69_1_1: int = __storage_get_extern(0x1111111111111111111111111111111111111111111111111111111111111111, {3, 69, ((0 + 1) + 1)});
-//     constraint (::u == {0x0000000000000000000000000000000000000000000000000000000000000000, 0});
-//     constraint (::u0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::u1 == 0);
-//     constraint (::t == {0x0000000000000000000000000000000000000000000000000000000000000000, {0, 0}});
-//     constraint (::t0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::t10 == 0);
-//     constraint (::t11 == 0);
-//     constraint (::w == {addr: 0x0000000000000000000000000000000000000000000000000000000000000000, inner: {0, 0}});
-//     constraint (::w0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::w10 == 0);
-//     constraint (::w11 == 0);
-//     constraint (::map_to_tuples_69 == {0x0000000000000000000000000000000000000000000000000000000000000000, {0, 0}});
-//     constraint (::map_to_tuples_69_0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::map_to_tuples_69_1_0 == 0);
-//     constraint (::map_to_tuples_69_1_1 == 0);
-//     constraint (::foo_u == {0x0000000000000000000000000000000000000000000000000000000000000000, 0});
-//     constraint (::foo_u0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::foo_u1 == 0);
-//     constraint (::foo_t == {0x0000000000000000000000000000000000000000000000000000000000000000, {0, 0}});
-//     constraint (::foo_t0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::foo_t10 == 0);
-//     constraint (::foo_t11 == 0);
-//     constraint (::foo_w == {addr: 0x0000000000000000000000000000000000000000000000000000000000000000, inner: {x: 0, 0}});
-//     constraint (::foo_w0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::foo_w10 == 0);
-//     constraint (::foo_w11 == 0);
-//     constraint (::foo_map_to_tuples_69 == {0x0000000000000000000000000000000000000000000000000000000000000000, {0, 0}});
-//     constraint (::foo_map_to_tuples_69_0 == 0x0000000000000000000000000000000000000000000000000000000000000000);
-//     constraint (::foo_map_to_tuples_69_1_0 == 0);
-//     constraint (::foo_map_to_tuples_69_1_1 == 0);
 //     constraint __eq_set(__mut_keys(), {0});
 // }
 // >>>
