+++ pintc/tests/tests.rs
@@ -55,9 +55,9 @@ fn run_tests(sub_dir: &str) -> anyhow::Result<()> {
             .and_then(|pred|
                 // Flatten the parsed intent and check the result.
                 flatten_and_check(pred, &test_data, &mut failed_tests, &path))
-            .and_then(|pred| 
+            .and_then(|pred|
                 // Optimise the flattened intent
-                optimise(pred, &test_data, &mut failed_tests, &path));
+                optimise(pred, &test_data));
 
         // Check the `json` ABI if a reference file exists.
         if let Some(program) = program {
@@ -240,29 +240,14 @@ fn flatten_and_check(
         .ok()
 }
 
-fn optimise(
-    pred: Contract,
-    test_data: &TestData,
-    failed_tests: &mut Vec<String>,
-    path: &Path,
-) -> Option<Contract> {
+fn optimise(pred: Contract, test_data: &TestData) -> Option<Contract> {
     let optimised = pred.optimise();
     if let Some(expected_optimised_str) = &test_data.optimised {
-        similar_asserts::assert_eq!(
-            expected_optimised_str.trim(),
-            format!("{optimised}").trim()
-        );
+        similar_asserts::assert_eq!(expected_optimised_str.trim(), format!("{optimised}").trim());
         Some(optimised)
     } else {
-        failed_tests.push(path.display().to_string());
-        println!(
-            "{} {}.",
-            "MISSING 'optimised' DIRECTIVE".red(),
-            path.display().to_string().cyan(),
-        );
         None
     }
-      
 }
 
 #[cfg(test)]
@@ -296,4 +281,5 @@ mod e2e {
     e2e_test!(root_types);
     e2e_test!(storage);
     e2e_test!(types);
+    e2e_test!(optimisations);
 }
